# be/docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A comprehensive donation and auction management system that transforms donated goods into meaningful impact through transparent auctions and direct donations to institutions in need.",
        "title": "Your Donate Rise API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "yourdonaterise-278016640112.asia-southeast2.run.app",
    "basePath": "/",
    "paths": {
...
        "/auction/sessions/{sessionID}/items/{itemID}/bid": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Place a bid on a specific auction item within an active session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Your Donate Rise API - Bidding"
                ],
                "summary": "Place bid on auction item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Auction Session ID",
                        "name": "sessionID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Auction Item ID",
                        "name": "itemID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Bid amount",
                        "name": "bid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BidDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bid placed successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid parameters or bid too low",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing token",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Auction session or item not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict - Invalid auction state",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auction/sessions/{sessionID}/items/{itemID}/highest-bid": {
            "get": {
                "description": "Retrieve the current highest bid for a specific auction item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Your Donate Rise API - Bidding"
                ],
                "summary": "Get highest bid for auction item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Auction Session ID",
                        "name": "sessionID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Auction Item ID",
                        "name": "itemID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Highest bid retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid session or item ID",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Auction not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auction/sessions/{sessionID}/items/{itemID}/sync": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Synchronize the highest bid from Redis cache to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Your Donate Rise API - Bidding"
                ],
                "summary": "Sync highest bid to database",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Auction Session ID",
                        "name": "sessionID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Auction Item ID",
                        "name": "itemID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Highest bid synced to database",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid session or item ID",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing token",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin access required",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.BidDTO": {
            "type": "object",
            "required": [
                "amount"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                }
            }
        },
...
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
        "utils.SuccessResponseData": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}